requirements.txt
PySimpleGUI
torch ~= 1.10.1
torchvision ~= 0.11.2
Pillow
numpy
opencv-python
